fixed collisions

+++ app/src/main/java/com/example/bullethellworld/views/Collidable.java
@@ -1,5 +1,5 @@
 package com.example.bullethellworld.views;
 
 public interface Collidable {
-    public boolean collides(float x0, float y0, float w, float h);
+    public boolean collides(float x0, float y0, int w, int h);
 }
