+++ app/src/main/java/com/example/bullethellworld/views/Frame.java
@@ -2,5 +2,5 @@ package com.example.bullethellworld.views;
 
 public interface Frame {
 
-    public Side collides(float x0, float y0, float w, float h);
+    public Side collides(float x0, float y0, int w, int h);
 }
